Move save_in to file_format.
